build: namespace the generated `version.h`
Namespaced the generated `version.h` with `zephyr` to prevent
potential conflict with other headers.

Introduce a temporary Kconfig `LEGACY_GENERATED_INCLUDE_PATH`
that is deprecated and enabled by default. This allows the
developers to continue the use of the old include paths for the
time being until it is removed. The Kconfig will generate a
build-time deprecation message.

Updated the includes path of in-tree sources accordingly.

Other generated headers will be worked on in the following commit

Signed-off-by: Yong Cong Sin <ycsin@meta.com>