posix: rename posix_features.h to sys/features.h
POSIX feature test macros are provided by the implementation to
the toolchain.

Although not part of the specification, the conventional header
file to use for POSIX features is, and it is used by glibc,
picolibc, and newlib, among many others.

<sys/features.h>

Rename our header posix_features.h to sys/features.h within the
common libc include path.

When POSIX features are selected via Kconfig, include sys/features.h via the
command line so that feature test macros are available to the application
when _POSIX_C_SOURCE is defined.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>