[WIP]
Trying to get the `_current` to use the `z_tls_current` var in the TLS.

The idea is to ALWAYS have the `tp` intialized (even before we reach main, in the '_dummy_thread'),
and `z_tls_current` configured at thread/stack creation before the thread is spawned, so that
`z_tls_current` always points to the right thread.

I'm trying to get the `_current` to use the `z_tls_current` var in the TLS, but haven't got any success so far.
The idea is to initialize the `z_tls_current` in `arch_tls_stack_setup()`, before the thread can be scheduled,
so that `z_tls_current` always point to the currently running thread.

One exception being the dummy_thread, which doesn't have a stack, but we can workaround it by creating
a dummy stack (or borrow from idle stack, if possible?)

How should I interpret the differences between `_current_cpu->current`, `_current` & `k_current_get()`?